Add begin/rescue with retry to try to reconnect to db
This is a known issue with octopus that is caused by trying
to switch between db connections. From what we can gather, during
the switch, the underlying database connections are not available
when the query is run. This code catches the "PG::ConnectionBad"
error and tries to run the #verify method to create the connections
to the database before retrying.

RESOURCES:
* https://github.com/thiagopradi/octopus/issues/457
* https://github.com/procore/octopus/commit/c015aacf60360e83738550dcbd1315d1cd383ae3